refactor(cobros): extraer calcularDiasMoraExactos a lib/mora-utils
Elimina duplicación de la función en cobros.ts, reportes-cartera.ts y
sync-casos-cobros.ts, centralizándola en un utility compartido.
